Move annotations into the server-api artifact
