+++ mcp-server-api/src/main/java/org/mcp_java/server/annotations/package-info.java
@@ -24,18 +24,18 @@
  *
  * <h2>Core Annotations</h2>
  * <ul>
- * <li>{@link org.mcp_java.annotations.McpServer @McpServer} - Marks a class as an MCP server component</li>
- * <li>{@link org.mcp_java.annotations.MetaField @MetaField} - Adds custom metadata fields to MCP definitions</li>
+ * <li>{@link org.mcp_java.server.annotations.McpServer @McpServer} - Marks a class as an MCP server component</li>
+ * <li>{@link org.mcp_java.server.annotations.MetaField @MetaField} - Adds custom metadata fields to MCP definitions</li>
  * </ul>
  *
  * <h2>Sub-packages</h2>
  * <ul>
- * <li>{@link org.mcp_java.annotations.tools} - Annotations for defining MCP tools</li>
- * <li>{@link org.mcp_java.annotations.resources} - Annotations for defining MCP resources</li>
- * <li>{@link org.mcp_java.annotations.prompts} - Annotations for defining MCP prompts</li>
- * <li>{@link org.mcp_java.annotations.completion} - Annotations for auto-completion functionality</li>
+ * <li>{@link org.mcp_java.server.tools.annotations} - Annotations for defining MCP tools</li>
+ * <li>{@link org.mcp_java.server.resources.annotations} - Annotations for defining MCP resources</li>
+ * <li>{@link org.mcp_java.server.prompts.annotations} - Annotations for defining MCP prompts</li>
+ * <li>{@link org.mcp_java.server.completion.annotations} - Annotations for auto-completion functionality</li>
  * </ul>
  *
  * @see <a href="https://modelcontextprotocol.io">Model Context Protocol</a>
  */
-package org.mcp_java.annotations;
+package org.mcp_java.server.annotations;
