+++ mcp-server-api/src/main/java/org/mcp_java/server/completion/annotations/CompleteResourceTemplate.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.mcp_java.annotations.completion;
+package org.mcp_java.server.completion.annotations;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -21,10 +21,12 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import org.mcp_java.server.completion.CompletionResult;
+
 /**
- * Marks a method as providing completion suggestions for a prompt argument.
+ * Marks a method as providing completion suggestions for a resource template URI expression.
  * <p>
- * The annotated method should return completion values for a specific prompt argument.
+ * The annotated method should return completion values for URI template variables.
  * The method must consume exactly one {@link String} argument representing the
  * partial value that needs to be completed.
  * </p>
@@ -34,30 +36,31 @@ import java.lang.annotation.Target;
  * <ul>
  * <li>{@code String} - single completion value</li>
  * <li>{@code List<String>} - multiple completion values</li>
- * <li>{@code CompletionResponse} - full completion response with metadata</li>
+ * <li>{@link CompletionResult} - full completion response with metadata</li>
  * </ul>
  * <p>
  * Example:
  * </p>
  * <pre>
- * &#64;CompletePrompt("myPrompt")
- * public List&lt;String&gt; completeMyPromptArg(&#64;CompleteArg String partialValue) {
- *     return List.of("option1", "option2", "option3");
+ * &#64;CompleteResourceTemplate("fileTemplate")
+ * public List&lt;String&gt; completeFilePath(&#64;CompleteArg String partialPath) {
+ *     // Return matching file paths
+ *     return findMatchingPaths(partialPath);
  * }
  * </pre>
  */
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface CompletePrompt {
+public @interface CompleteResourceTemplate {
 
     /**
-     * The name of the prompt this completion method is for.
+     * The name of the resource template this completion method is for.
      * <p>
-     * This must reference an existing {@code @Prompt} annotation's name.
+     * This must reference an existing {@code @ResourceTemplate} annotation's name.
      * </p>
      *
-     * @return the prompt name
+     * @return the resource template name
      */
     String value();
 }
