+++ mcp-server-api/src/main/java/org/mcp_java/server/prompts/annotations/package-info.java
@@ -18,10 +18,10 @@
  * Annotations for defining MCP prompts.
  * <p>
  * Prompts are reusable message templates that can be used by MCP clients.
- * Use {@link org.mcp_java.annotations.prompts.Prompt @Prompt} to mark methods as prompts
- * and {@link org.mcp_java.annotations.prompts.PromptArg @PromptArg} to configure their arguments.
+ * Use {@link org.mcp_java.server.prompts.annotations.Prompt @Prompt} to mark methods as prompts
+ * and {@link org.mcp_java.server.prompts.annotations.PromptArg @PromptArg} to configure their arguments.
  * </p>
  *
  * @see <a href="https://modelcontextprotocol.io/specification/2025-11-25/server/prompts">MCP Specification - Prompts</a>
  */
-package org.mcp_java.annotations.prompts;
+package org.mcp_java.server.prompts.annotations;
