+++ mcp-server-api/src/main/java/org/mcp_java/server/tools/Tools.java
@@ -16,11 +16,12 @@
 package org.mcp_java.server.tools;
 
 import org.mcp_java.server.content.TextContent;
+import org.mcp_java.server.tools.annotations.Tool;
 
 /**
  * Provides access to builders for Tool responses
  * <p>
- * This class can be injected as a parameter to a {@code @Tool}-annotated method, or using framework-specific mechanisms.
+ * This class can be injected as a parameter to a {@link Tool}-annotated method, or using framework-specific mechanisms.
  */
 public interface Tools {
